+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/Templates/ValidationServiceInterface/ValidationServiceInterfaceTemplateRegistration.cs
@@ -32,7 +32,8 @@ namespace Intent.Modules.Application.FluentValidation.Dtos.Templates.ValidationS
                         dtoTemplateId: TemplateRoles.Application.Contracts.Dto,
                         dtoValidatorTemplateId: TemplateRoles.Application.Validation.Dto,
                         uniqueConstraintValidationEnabled: application.Settings.GetFluentValidationApplicationLayer().UniqueConstraintValidation().IsDefaultEnabled(),
-                        customValidationEnabled: true)))
+                        customValidationEnabled: true,
+                        associationedElements: x.InternalElement?.AssociatedElements)))
             {
                 AbortRegistration(); // Need cleaner, more obvious way, to do this
                 return;