+++ Modules/Intent.Modules.Application.MediatR.FluentValidation/Templates/QueryValidator/QueryValidatorTemplatePartial.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Collections.Generic;
 using Intent.Engine;
+using Intent.Metadata.Models;
 using Intent.Modelers.Services.Api;
 using Intent.Modelers.Services.CQRS.Api;
 using Intent.Modules.Application.FluentValidation.Settings;
@@ -24,7 +26,7 @@ namespace Intent.Modules.Application.MediatR.FluentValidation.Templates.QueryVal
         public const string TemplateId = "Intent.Application.MediatR.FluentValidation.QueryValidator";
 
         [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
-        public QueryValidatorTemplate(IOutputTarget outputTarget, QueryModel model)
+        public QueryValidatorTemplate(IOutputTarget outputTarget, QueryModel model, IEnumerable<IAssociationEnd> associationedElements)
             : base(
                 templateId: TemplateId,
                 outputTarget: outputTarget,
@@ -37,6 +39,7 @@ namespace Intent.Modules.Application.MediatR.FluentValidation.Templates.QueryVal
                 uniqueConstraintValidationEnabled: outputTarget.ExecutionContext.Settings.GetFluentValidationApplicationLayer().UniqueConstraintValidation().IsDefaultEnabled(),
                 repositoryInjectionEnabled: true,
                 customValidationEnabled: true,
+                associationedElements: associationedElements,
                 additionalFolders: outputTarget.ExecutionContext.Settings.GetCQRSSettings().ConsolidateCommandQueryAssociatedFilesIntoSingleFile()
                     ? Array.Empty<string>()
                     : new[] { model.GetConceptName() })
@@ -44,7 +47,7 @@ namespace Intent.Modules.Application.MediatR.FluentValidation.Templates.QueryVal
             FulfillsRole(TemplateRoles.Application.Validation.Query);
         }
 
-        public static void Configure(CSharpTemplateBase<QueryModel> template)
+        public static void Configure(CSharpTemplateBase<QueryModel> template, IEnumerable<IAssociationEnd> associationedElements)
         {
             template.ConfigureForValidation(
                 dtoModel: new DTOModel(template.Model.InternalElement),
@@ -55,7 +58,8 @@ namespace Intent.Modules.Application.MediatR.FluentValidation.Templates.QueryVal
                 validatorProviderInterfaceTemplateId: "Application.Common.ValidatorProviderInterface",
                 uniqueConstraintValidationEnabled: template.ExecutionContext.Settings.GetFluentValidationApplicationLayer().UniqueConstraintValidation().IsDefaultEnabled(),
                 repositoryInjectionEnabled: true,
-                customValidationEnabled: true);
+                customValidationEnabled: true,
+                associationedElements: associationedElements);
         }
 
         public override bool CanRunTemplate()